Redirect stdout *and* stderr to /dev/null
The primary change is to also redirect stderr, which was previously
missed because I has smol brain.

Additionally replace "> /dev/null/" with "1>/dev/null", which solves the
previous problem where it looked a bit like we were checking whether
"type {command}" was greater than "/dev/null". And makes it clearer that
stream "1" is being redirected.