Compile and run programs in separate steps
The only real benefit of using the shorthand "run" options previously
used in the runner's for C#, Go and Rust are that they are quick to type
out, an advantage that is lost considering they are only used in
scripts.

Additionally remove redundant checking for position of first argument in
C# fizzbuzz and is_prime implementations.